refactor(subagent): move discord-sink to transports/ (#485)
Discord-specific output formatting belongs with the transport layer,
not the subagent subsystem. Moved discord-sink.ts to
transports/discord-subagent-sink.ts. Re-exports preserved in
subagent/index.ts for backward compat.
